Handle mapping failure from get_ids.
Signed-off-by: Kevin Coffman <kwc@citi.umich.edu>

Temporary patch to do default mapping if we get an error while trying to
map a gss principal to the appropriate uid/gid.  This currently returns
hardcoded values.  This may be correct, or we may need to try and figure
out the correct values to match the anonuid/anongid for the export.