Restore use of un-connected socket for UDP mount requests.
When connecting to an NFSv4 server we need to find out IP address
as it would be seen by the server, to register an address for
callbacks.
This is most easily done by connecting the socket to the
servers address and then getting the address of our endpoint.

However with a connected UDP socket, replies that come from a
different IP address - as can happen with non-Linux multi-homed
servers - will be rejected.

So if we connected our UDP socket, we need to be sure to
disconnect it before using it.

This patch adds an option to get_socket to say if we want it
connected or not and, in the case where we do, we disconnect
a UDP socket after the connection information has been used.

Also clean up the error handling in clnt_ping which was getting
clumsy.