If -o sec= is not specified for mount, allow either AUTH_SYS or AUTH_NONE.
Solaris servers, when asked to share a filesystem with an anon-uid, will
report the only available authentication style as AUTH_NONE in the
reply from mountd (even though they actually accept AUTH_SYS and simply
ignore the credentiuals).
So if no sec= is specified we should really accept anything that can easily
be handled.  ie. AUTH_SYS or AUTH_NONE.