Automatically start statd when mounting an nfs filesystem.
If statd is not running (/var/run/rpc.statd.pid) when an nfs filesystem
is mounted (v2 or v3, with remote locking enabled), and if
/usr/sbin/start-statd (or other program specified at config time) is
present, then run that program to start statd.

This means that statd does not need to be running "just in case".
It only needs to be started at boot time if the nfs server is
started.