Correctly handle "user" and "users" mount options.
If "user" or "users" is given, then allow mount.nfs to be run
by a non-root user providing that the mountpoint, filesystem, and options
exactly match what is found in fstab.

For "user", record the user name in mtab so they can unmount the
filesystem later.

Also alwasys ignore auto, owner, group and their negations as well
as "_netdev", "comment" and "loop".