Handle -o remount better
On -o remount, we need to update the entry in mtab rather than
add a new one.  update_mtab does this so use that.
However it might free some strings that shouldn't be freed, so
stop it from calling free - the program will exit soon anyway
so no exit is needed.