statd - keep persistent state in sm/* files.
If statd dies and is restarted, it forgets what peers the kernel
is interested in monitoring, and so will not forward NOTIFY
requests properly.

With this patch the required information is recorded in the files
in /var/lib/nfs/sm/* so that a kill/restart does what you might
hope.

Signed-off-by: Neil Brown <neilb@suse.de>