statd - only unregister/register once.
The for loop that restarts on SIGUSR or simu_reboot currently includes
several once-only things, that are probably best taken out of the loop.

We also take the unregister/register out of the loop as if statd does
drop privileges, then the second register won't use a privileged port
properly.

On the whole, cleaner code.

Signed-off-by: Neil Brown <neilb@suse.de>