mountd - improve checks and error messages for export failure.
If an attempt is made to export a non-(dir|file), just ignore it.
This should get caught by exportfs.

If an attempt is made to export a non-exportable filesystem, report
an error.  Hopefully exportfs can trap some these as well, but
catching them in mountd as well is good.

Signed-off-by: Neil Brown <neilb@suse.de>