Be more cautious about use for privilege ports (<1024).
Ports < 1024 are a scarce resource and should not be used
carelessly.  Technically they should be not used at all without
registration with IANA, but sometimes we need them despite that.

So: for the socket that RPC services listen on, don't use a <1024 port
by default.  There is no need.

For sockets that we send messages on, that are long-lived, and that might
need to appear 'privileged', avoid using a number that is registered in
/etc/services if possible.