mount.nfs: Always preset nfs_mount_version
nfs_mount_version is a global integer that is set based on a guess
about which nfs_mount_data version is appropriate for the kernel we're
running on.

Make it always available and have the correct value before calling mount
and unmount so they don't have to worry about setting it themselves.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Neil Brown <neilb@suse.de>