libnfs.a: eliminate another dependency on a global variable
The file support/nfs/fstab.c, which is linked into libnfs.a, depends on the
global variable "verbose."  This variable is defined and used only in the
mount command, and the functions in fstab.c are used only by the mount
command.

Move fstab.c and support/include/fstab.h to utils/mount.  This file
placement is also consistent with at least one other mount helper,
mount.ocfs2.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Neil Brown <neilb@suse.de>