nfsstat: Add -S/--since flag.
	Read statistics from the file given with -S/--since and display
the difference between those and the current statistics.  Valid stat files
are those in the form of /proc/net/rpc/nfs, /proc/net/rpc/nfsd, or any
"pretty" output from nfsstat itself.  Statistics that are missing from a
"pretty" stat file are treated as zeroes.

	Similar to Neil Brown's suggestion, one might use this in
conjunction with watch(1) like this:

$ watch "nfsstat --since /tmp/stats; nfsstat >/tmp/stats"

Signed-off-by: David M. Richter <richterd@citi.umich.edu>
Signed-off-by: Neil Brown <neilb@suse.de>