text-based mount.nfs: Fix mounthost= processing
The 'mounthost=' option names a host where the mountd service is running.
The option is used to direct clients to use a different host for the mountd
procotol than the host where the NFS service is running.

The nfs(5) man page shows that the 'mounthost=' option takes a name, not
an address.  The kernel's text-based mount option parsing logic expects an
IPv4 address.  This is necessary because the kernel cannot itself resolve
hostnames to addresses.

Resolve the hostname and pass in a new mount option that contains the
resolved address, 'mountaddr=', to the kernel.

This requires a patch to the kernel to recognize the new 'mountaddr='
option, and to change the 'mounthost=' parsing logic to treat the value of
this option as a simple string.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Neil Brown <neilb@suse.de>