Fix version fallback for unmount.
Previously, if the mtab record didn't mention a version, unmount
would assume a v3 umount and send an UNMOUNT request accordingly.
This is wrong.

So remove the 'v3' assumption, and allow probe_port to continue when
it gets a version number mis-match.

Also there was some overloading of the meaning of pm_vers==0 relating
to v4 mounts.  As do_nfs_umount is never called for v4, rename it to
do_nfs_umount23, and remove v4 handling from there and from
nfs_call_umount.

Signed-off-by: Neil Brown <neilb@suse.de>