Lots of parts of nfs-utils already depend on getaddrinfo(3).
We could find each instance where getaddrinfo(3) is invoked, wrap it with
'#ifdef HAVE_GETADDRINFO', and provide equivalent logic without it, but that's
a whole lot of work... and no-one has complained about this so far.

So as a clean-up, let's simply add a hard dependency for it in configure.ac,
and call it a day.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>