showmount command: clean up error returns from connect_nb()
Clean up connect_nb() in the showmount command.

Sometimes it returns -1 on error, and sometimes a negative errno.  On error,
it should always return one of these or the other, not both.

Similar functions in other parts of nfs-utils return -1 on error, and set
errno; so let's do that here too.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>