rpc.statd: Clean up: replace "if (!(foo = rtnl))".
Static code checkers flag this kind of thing because it's easy to
confuse with "if (!(foo == rtnl))".  In one of these cases, the
combination of evaluation and assignment isn't even necessary.

While we are in the neighborhood, remove an extra argument to note() that is
not called for in the passed-in format string.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>