Add AF_INET6-capable API to acquire an RPC	CLIENT *
Provide a simple interface that any component of nfs-utils can use to acquire
an RPC CLIENT *.  This is an AF_INET6-enabled API, and can also handle
PF_LOCAL sockets if libtirpc is present on the system.

When libtirpc is not available, legacy RPC services will be used instead,
and an attempt to connect to an AF_INET6 address will fail.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>